Move replays to `common`
